Remove parallel_xexec function from thread helpers
syncthreads_barrier was being passed as syncthreads function by
value (copy). This results in all threads having it's own copy
of synchtreads_barrier object. In consequence, assigning new
latch to syncthreads_barrier->current_latch was done on a local
copy of syncthreads_barrier, instead of on a shared object. This
resulted in a deadlock in case there was more than one call to
synchtreads.

Fix this, by requiring user to manually create syncthreads_barrier
and prohibit copying it.