Get rid region pamarmeter from pmemstream_region_iterator_next
Instead of that expose pmemstream_region_iterator structure to the user

This simplify usage of region_iterators. Now if user wants store
iterators in some structure, doesn't need to store stream and region
objects.

In  this implementation new iterator always points to offset 0 if it
was allocated or INVALID_REGION otherwise.